+++ visualization/app/codeCharta/core/url/url.service.ts
@@ -1,5 +1,11 @@
 "use strict";
 import {ILocationService, IHttpService, IHttpResponse} from "angular";
+import {CodeMap} from "../data/model/CodeMap";
+
+export interface NameDataPair {
+    name: string;
+    data: Object;
+}
 
 /**
  * This service offers an application specific url API.
@@ -54,15 +60,30 @@ export class UrlService {
     }
 
     /**
-     * returns the files content specified in the 'file' url parameter
+     * returns the files contents specified in the 'file' url parameter
      * @returns {Promise} which returns the files content on resolution
      */
-    public getFileDataFromQueryParam(): Promise<Object> {
+    public getFileDataFromQueryParam(): Promise<NameDataPair[]> {
 
-        return new Promise((resolve, reject) => {
-            let param = this.getParam("file");
-            this.getFileDataFromFile(param).then(resolve, reject);
+        let fileNames = this.$location.search().file;
+
+        if(!fileNames) {
+            fileNames = [];
+        }
+
+        if(fileNames.push === undefined) {
+            fileNames = [fileNames];
+        }
+
+        let fileReadingTasks = [];
+
+        fileNames.forEach((fileName)=>{
+            fileReadingTasks.push(new Promise((resolve, reject) => {
+                this.getFileDataFromFile(fileName).then(resolve, reject);
+            }));
         });
+
+        return Promise.all(fileReadingTasks);
 
     }
 
@@ -70,7 +91,7 @@ export class UrlService {
      * returns the files content specified in the 'file' parameter
      * @returns {Promise} which returns the files content on resolution
      */
-    public getFileDataFromFile(file: string): Promise<Object> {
+    public getFileDataFromFile(file: string): Promise<NameDataPair> {
 
         return new Promise((resolve, reject) => {
 
@@ -79,7 +100,7 @@ export class UrlService {
                     function (response: IHttpResponse<Object>) {
                         if (response.status === UrlService.OK_CODE) {
                             Object.assign(response.data, {fileName: file});
-                            resolve(response.data);
+                            resolve({name: file, data:response.data});
                         } else {
                             reject();
                         }